% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBeds.R
\name{getBeds}
\alias{getBeds}
\title{COVID-19 Hospital statistics}
\usage{
getBeds(codes = NULL, dates = NULL, indicator = NULL)
}
\arguments{
\item{codes}{List or string of country codes}

\item{dates}{List of dates for which you want
the data (you can use the getDays function)}

\item{indicator}{List of string of the indicator you want.
You can have :
Daily ICU occupancy (ICU)
Weekly new ICU admissions per 100k (ICUPER)
Daily hospital occupancy (HOSP)
Weekly new hospital admissions per 100k (HOSPPER)}
}
\value{
A data frame containing the data for the indicator selected for each
countries mentioned
}
\description{
COVID-19 Hospital statistics
}
\examples{
getBeds(codes = c("BE", "FR"), dates = c("2021-01-03", "2021-01-04"), indicator = "icu")
}
