% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCovid.R
\name{getCovid}
\alias{getCovid}
\title{COVID-19 statistics}
\usage{
getCovid(codes = NULL, dates = NULL)
}
\arguments{
\item{codes}{List or string of country codes}

\item{dates}{List of dates for which you want
the data (you can use the getDays function)}
}
\value{
A data frame containing the number of cases and deaths for each
countries mentioned as well as the latest population data
}
\description{
COVID-19 statistics
}
\examples{
getCovid()
getCovid(codes = "FR")
getCovid(dates = "2021-04-01")
getCovid(codes = c("AT", "FR"),
dates = getDays(startDate = "2021-04-01", endDate = "2021-04-04"))
}
