% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDays.R
\name{getDays}
\alias{getDays}
\title{List all days between two dates}
\usage{
getDays(startDate, endDate)
}
\arguments{
\item{startDate}{A string}

\item{endDate}{A string}
}
\value{
A vector containing the list of all days between
the startDate and the endDate
}
\description{
List all days between two dates
}
\examples{
getDays(startDate = "2021-04-01", endDate = "2021-04-04")
}
