% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeasures.R
\name{getMeasures}
\alias{getMeasures}
\title{COVID-19 Political Measures}
\usage{
getMeasures(codes = NULL, measure = NULL, dates = NULL)
}
\arguments{
\item{codes}{List or string of country codes}

\item{measure}{List or string for the measure you want data on.
You can find more information on the measures in the data dictionary of the
ECDC available here :
https://www.ecdc.europa.eu/en/publications-data/download-data-response-measures-covid-19}

\item{dates}{List of dates for which you want
the data (you can use the getDays function)}
}
\value{
A data frame containing the data for the political measures
against COVID-19 for each countries mentioned
}
\description{
COVID-19 Political Measures
}
\examples{
getMeasures(codes = c("CH", "AT", "FR"), measure = "MassGatherAll", dates = "2020-03-01")
}
