% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTests.R
\name{getTests}
\alias{getTests}
\title{COVID-19 Tests Statistics}
\usage{
getTests(codes = NULL, dates = NULL)
}
\arguments{
\item{codes}{List or string of country codes}

\item{dates}{List of dates for which you want
the data (you can use the getDays function)}
}
\value{
A data frame containing the data on testing for each
countries mentioned
}
\description{
COVID-19 Tests Statistics
}
\examples{
getTests(codes = c("AT", "FR"), dates = getDays(startDate = "2020-10-01", endDate = "2021-01-01"))
}
