% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVaccines.R
\name{getVaccines}
\alias{getVaccines}
\title{COVID-19 Vaccines Statistics}
\usage{
getVaccines(codes = NULL, dates = NULL, vac = NULL, target = NULL)
}
\arguments{
\item{codes}{List or string of country codes}

\item{dates}{List of dates for which you want
the data (you can use the getDays function)}

\item{vac}{A string as identifier of the vaccine you want data on.
You can find more information on the measures in the data dictionary of the
ECDC available here :
https://www.ecdc.europa.eu/en/publications-data/covid-19-testing}

\item{target}{A string as identifier of the target group for the vaccine
you want data on.
You can find more information on the measures in the data dictionary of the
ECDC available here :
https://www.ecdc.europa.eu/en/publications-data/covid-19-testing}
}
\value{
A data frame containing the data on the vaccines
against COVID-19 for each countries mentioned
}
\description{
COVID-19 Vaccines Statistics
}
\examples{
getVaccines(codes = c("DE", "FR"), dates = "2021-01-01", vac = "AZ", target = "ALL")
}
