% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariants.R
\name{getVariants}
\alias{getVariants}
\title{COVID-19 Variants Statistics}
\usage{
getVariants(codes = NULL, dates = NULL, var = NULL)
}
\arguments{
\item{codes}{List or string of country codes}

\item{dates}{List of dates for which you want
the data (you can use the getDays function)}

\item{var}{String as identifier of the variant you want data on.
You can find more information on the measures in the data dictionary of the
ECDC available here :
www.ecdc.europa.eu/en/publications-data/data-virus-variants-covid-19-eueea}
}
\value{
A data frame containing the data for the political measures
against COVID-19 for each countries mentioned
}
\description{
COVID-19 Variants Statistics
}
\examples{
getVariants(codes = c("AT", "DE"), dates = getDays(startDate = "2020-12-01",
endDate = "2021-01-01"), var = "Other")
}
