% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeeks.R
\name{getWeeks}
\alias{getWeeks}
\title{Get the ISO week of dates with or without a W}
\usage{
getWeeks(dates, sym = TRUE)
}
\arguments{
\item{dates}{A string}

\item{sym}{A logical}
}
\value{
A vector containing the list of the ISO weeks of dates
}
\description{
Get the ISO week of dates with or without a W
}
\examples{
getWeeks(dates = c("2021-04-01", "2021-04-04", "2021-05-06"), sym = TRUE)
getWeeks(dates = c("1962-03-26", "1966-10-06", "1998-07-21", "2000-03-08"), sym = FALSE)
}
